web-backend/requirements.txt
flask
boto3
pillow